feat: common UI 생성
- reset.css 세팅 및 사용하지 않는 css 제거
- Layout.jsx를 이용해서 고정된 header/footer가 노출되도록 처리
- header에 nav 영역 추가
- 디렉토리별 파일 구분하여 용도가 분명하도록 함
- Layout main 컴포넌트를 전달하는 구조로 변경하면서 기존 App.jsx에 있던 CheckListGroup을 컴포넌트로 분리

#18
