+++ frontend/src/assets/css/Layout.css
@@ -1,0 +1,65 @@
+:root {
+  --primary-color: #74c0fc;
+}
+
+body {
+  margin: 0;
+  font-family: 'Segoe UI', sans-serif;
+  background-color: #f9f9f9;
+}
+
+.layout {
+  min-height: 100vh;
+  display: flex;
+  flex-direction: column;
+}
+
+.header {
+  background-color: white;
+  border-bottom: 1px solid #e5e5e5;
+  padding: 16px 32px;
+  align-items: center;
+}
+
+.header-inner {
+  max-width: 1200px;
+  margin: 0 auto;
+  display: flex;
+  justify-content: space-between;
+}
+
+.logo {
+  font-size: 20px;
+  font-weight: bold;
+  color: var(--primary-color);
+}
+
+.header-right .menu-item {
+  margin-left: 24px;
+  color: #333;
+  text-decoration: none;
+  font-weight: 500;
+  transition: color 0.2s ease;
+}
+
+.header-right .menu-item:hover {
+  color: var(--primary-color);
+}
+
+.main-wrapper {
+  flex: 1;
+  padding: 32px;
+}
+
+.main {
+    width: 1200px;
+    margin: 0 auto;
+}
+
+.footer {
+  background-color: #f0f0f0;
+  padding: 16px 32px;
+  text-align: center;
+  color: #888;
+  font-size: 14px;
+}
