+++ frontend/src/layouts/Footer.jsx
@@ -1,0 +1,7 @@
+export default function Footer() {
+    return (
+        <footer className="footer">
+            <p>© 2025 check-tracker. 좋은 습관, 좋은 하루.</p>
+        </footer>
+    );
+}
